+++ game/commander/tasks/compound/nextaction.py
@@ -8,6 +8,7 @@ from game.commander.tasks.compound.attackairinfrastructure import (
 )
 from game.commander.tasks.compound.attackbattlepositions import AttackBattlePositions
 from game.commander.tasks.compound.attackbuildings import AttackBuildings
+from game.commander.tasks.compound.attackmotorpools import AttackMotorpools
 from game.commander.tasks.compound.attackships import AttackShips
 from game.commander.tasks.compound.capturebases import CaptureBases
 from game.commander.tasks.compound.defendbases import DefendBases
@@ -65,6 +66,7 @@ class PlanNextAction(CompoundTask[TheaterState]):
             self.aircraft_cold_start
         ),
         "AttackBuildings": lambda self: AttackBuildings(),
+        "AttackMotorpools": lambda self: AttackMotorpools(),
         "AttackShips": lambda self: AttackShips(),
         "DegradeIads": lambda self: DegradeIads(),
     }